Move some types from tempTypes.ts to dedicated types files
