+++ utils/src/treev2/quickPickWizard/quickPickAzureResource/AzureResourceQuickPickWizardContext.ts
@@ -3,8 +3,9 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { ResourceGroupsItem } from "../../../../hostapi.v2";
 import { QuickPickWizardContext } from "../QuickPickWizardContext";
-import { ApplicationResource, ApplicationSubscription, ResourceGroupsItem } from "./tempTypes";
+import { ApplicationResource, ApplicationSubscription } from "./tempTypes";
 
 export interface AzureResourceQuickPickWizardContext extends QuickPickWizardContext<ResourceGroupsItem> {
     subscription: ApplicationSubscription | undefined;
