+++ utils/src/treev2/quickPickWizard/quickPickAzureResource/QuickPickAzureResourceGroupStep.ts
@@ -3,11 +3,10 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { ResourceGroupsItem, ContextValueFilterableTreeNode } from "../../../../hostapi.v2";
 import * as types from "../../../../index";
-import { ContextValueFilterableTreeNode } from "../ContextValueQuickPickStep";
 import { GenericQuickPickOptions, GenericQuickPickStep } from "../GenericQuickPickStep";
 import { AzureResourceQuickPickWizardContext } from "./AzureResourceQuickPickWizardContext";
-import { ResourceGroupsItem } from "./tempTypes";
 
 // TODO: implement this for picking resource group
 // The resource group may NOT be the grouping method used in the tree
