+++ utils/src/treev2/quickPickWizard/quickPickAzureResource/QuickPickGroupStep.ts
@@ -7,7 +7,8 @@ import * as types from '../../../../index';
 import * as vscode from 'vscode';
 import { GenericQuickPickStep, SkipIfOneQuickPickOptions } from '../GenericQuickPickStep';
 import { AzureResourceQuickPickWizardContext } from './AzureResourceQuickPickWizardContext';
-import { GroupingItem, ResourceGroupsItem } from './tempTypes';
+import { GroupingItem } from './tempTypes';
+import { ResourceGroupsItem } from '../../../../hostapi.v2';
 
 interface GroupQuickPickOptions extends SkipIfOneQuickPickOptions {
     groupType: types.AzExtResourceType;
